Module snes::instructions

All CPU instructions are implemented here

Many of the instructions are similar with different address modes, and can operate on 8 or 16 bit values depending on status flag settins. Therefor I decided to generate as many as I could automatically. The generated instructions use their respective address modes to aquire a target address. And they determine the param size based on the flags set in the State.

Functions

adc_6d

adc_6f

adc_7d

adc_7f

adc_61

adc_63

adc_65

adc_67

adc_69

adc_71

adc_72

adc_73

adc_75

adc_77

adc_79

and_2d

and_2f

and_3d

and_3f

and_21

and_23

and_25

and_27

and_29

and_31

and_32

and_33

and_35

and_37

and_39

asl_0a

asl_0e

asl_1e

asl_06

asl_16

bcc_90

bcs_b0

beq_f0

bit_2c

bit_3c

bit_24

bit_34

bit_89

bmi_30

bne_d0

bpl_10

bra_80

brk_00

brl_82

bvc_50

bvs_70

clc_18

cld_d8

cli_58

clv_b8

cmp_c1

cmp_c3

cmp_c5

cmp_c7

cmp_c9

cmp_cd

cmp_cf

cmp_d1

cmp_d2

cmp_d3

cmp_d5

cmp_d7

cmp_d9

cmp_dd

cmp_df

cop_02

cpx_e0

cpx_e4

cpx_ec

cpy_c0

cpy_c4

cpy_cc

dec_3a

dec_c6

dec_ce

dec_d6

dec_de

dex_ca

dey_88

eor_4d

eor_4f

eor_5d

eor_5f

eor_41

eor_43

eor_45

eor_47

eor_49

eor_51

eor_52

eor_53

eor_55

eor_57

eor_59

inc_1a

inc_e6

inc_ee

inc_f6

inc_fe

inx_e8

iny_c8

jmp_4c

jmp_5c

jmp_6c

jmp_7c

jmp_dc

jsr_20

jsr_22

jsr_fc

lda_a1

lda_a3

lda_a5

lda_a7

lda_a9

lda_ad

lda_af

lda_b1

lda_b2

lda_b3

lda_b5

lda_b7

lda_b9

lda_bd

lda_bf

ldx_a2

ldx_a6

ldx_ae

ldx_b6

ldx_be

ldy_a0

ldy_a4

ldy_ac

ldy_b4

ldy_bc

lsr_4a

lsr_4e

lsr_5e

lsr_46

lsr_56

mvn_54

mvp_44

nop_ea

ora_0d

ora_0f

ora_01

ora_1d

ora_1f

ora_03

ora_05

ora_07

ora_09

ora_11

ora_12

ora_13

ora_15

ora_17

ora_19

pea_f4

pei_d4

per_62

pha_48

phb_8b

phd_0b

phk_4b

php_08

phx_da

phy_5a

pla_68

plb_ab

pld_2b

plp_28

plx_fa

ply_7a

rep_c2

rol_2a

rol_2e

rol_3e

rol_26

rol_36

ror_6a

ror_6e

ror_7e

ror_66

ror_76

rti_40

rtl_6b

rts_60

sbc_e1

sbc_e3

sbc_e5

sbc_e7

sbc_e9

sbc_ed

sbc_ef

sbc_f1

sbc_f2

sbc_f3

sbc_f5

sbc_f7

sbc_f9

sbc_fd

sbc_ff

sec_38

sed_f8

sei_78

sep_e2

sta_8d

sta_8f

sta_9d

sta_9f

sta_81

sta_83

sta_85

sta_87

sta_91

sta_92

sta_93

sta_95

sta_97

sta_99

stp_db

stx_8e

stx_86

stx_96

sty_8c

sty_84

sty_94

stz_9c

stz_9e

stz_64

stz_74

tax_aa

tay_a8

tcd_5b

tcs_1b

tdc_7b

trb_1c

trb_14

tsb_0c

tsb_04

tsc_3b

tsx_ba

txa_8a

txs_9a

txy_9b

tya_98

tyx_bb

wai_cb

wdm_42

xba_eb

xce_fb